made corrections
