+++ src/assets/scss/pages/_talent.scss
@@ -1,17 +1,17 @@
-@import "../components/global/button";
-
-.title {
-  font-family: Rubik;
-  font-style: normal;
-  font-weight: normal;
-  font-size: 24px;
-  text-align: center;
-  color: #1f1f1f;
-  margin-bottom: 4rem;
-}
-.button__container {
-  width: 50%;
-  margin: auto;
-  text-align: center;
-  padding: 5rem 0;
+.talent {
+  &__title {
+    font-family: Rubik;
+    font-style: normal;
+    font-weight: normal;
+    font-size: 24px;
+    text-align: center;
+    color: #1f1f1f;
+    margin-bottom: 4rem;
+  }
+  &__button__container {
+    width: 50%;
+    margin: auto;
+    text-align: center;
+    padding: 5rem 0;
+  }
 }
