+++ src/data/TalentData.js
@@ -1,0 +1,57 @@
+const data = [
+  {
+    id: 1,
+    image:
+      "https://res.cloudinary.com/gfellah45/image/upload/v1601374966/img/pictures_tdnnso.png",
+    name: "Meyene Udoh",
+    title: "Software Developer",
+    socials: [
+      {
+        twitter: "https://twitter/meyeneudo",
+        github: "https://facebook/meyeneudoh",
+      },
+    ],
+  },
+  {
+    id: 2,
+    image:
+      "https://res.cloudinary.com/gfellah45/image/upload/v1601374966/img/pictures_tdnnso.png",
+    name: "Okoye CB",
+    title: "Software Developer",
+    socials: [{ twitter: "twitter/meyeneudo", github: "facebook/meyeneudoh" }],
+  },
+  {
+    id: 3,
+    image:
+      "https://res.cloudinary.com/gfellah45/image/upload/v1601374966/img/pictures_tdnnso.png",
+    name: "Okoye CB",
+    title: "Software Developer",
+    socials: [{ twitter: "twitter/meyeneudo", github: "facebook/meyeneudoh" }],
+  },
+  {
+    id: 4,
+    image:
+      "https://res.cloudinary.com/gfellah45/image/upload/v1601374966/img/pictures_tdnnso.png",
+    name: "Okoye CB",
+    title: "Software Developer",
+    socials: [{ twitter: "twitter/meyeneudo", github: "facebook/meyeneudoh" }],
+  },
+  {
+    id: 5,
+    image:
+      "https://res.cloudinary.com/gfellah45/image/upload/v1601374966/img/pictures_tdnnso.png",
+    name: "Okoye CB",
+    title: "Software Developer",
+    socials: [{ twitter: "twitter/meyeneudo", github: "facebook/meyeneudoh" }],
+  },
+  {
+    id: 6,
+    image:
+      "https://res.cloudinary.com/gfellah45/image/upload/v1601374966/img/pictures_tdnnso.png",
+    name: "Okoye CB",
+    title: "Software Developer",
+    socials: [{ twitter: "twitter/meyeneudo", github: "facebook/meyeneudoh" }],
+  },
+]
+
+export default data
