+++ src/pages/talent.jsx
@@ -4,81 +4,16 @@ import CandidateCard from "../components/talent/CandidateCard"
 import Hero from "../components/Hero"
 import ArrowForward from "@material-ui/icons/ArrowForward"
 import ArrowBack from "@material-ui/icons/ArrowBack"
+import data from "../data/TalentData"
 
 function Talent() {
-  // dummy data
-  const data = [
-    {
-      id: 1,
-      image:
-        "https://res.cloudinary.com/gfellah45/image/upload/v1601374966/img/pictures_tdnnso.png",
-      name: "Meyene Udoh",
-      title: "Software Developer",
-      socials: [
-        {
-          twitter: "https://twitter/meyeneudo",
-          github: "https://facebook/meyeneudoh",
-        },
-      ],
-    },
-    {
-      id: 2,
-      image:
-        "https://res.cloudinary.com/gfellah45/image/upload/v1601374966/img/pictures_tdnnso.png",
-      name: "Okoye CB",
-      title: "Software Developer",
-      socials: [
-        { twitter: "twitter/meyeneudo", github: "facebook/meyeneudoh" },
-      ],
-    },
-    {
-      id: 3,
-      image:
-        "https://res.cloudinary.com/gfellah45/image/upload/v1601374966/img/pictures_tdnnso.png",
-      name: "Okoye CB",
-      title: "Software Developer",
-      socials: [
-        { twitter: "twitter/meyeneudo", github: "facebook/meyeneudoh" },
-      ],
-    },
-    {
-      id: 4,
-      image:
-        "https://res.cloudinary.com/gfellah45/image/upload/v1601374966/img/pictures_tdnnso.png",
-      name: "Okoye CB",
-      title: "Software Developer",
-      socials: [
-        { twitter: "twitter/meyeneudo", github: "facebook/meyeneudoh" },
-      ],
-    },
-    {
-      id: 5,
-      image:
-        "https://res.cloudinary.com/gfellah45/image/upload/v1601374966/img/pictures_tdnnso.png",
-      name: "Okoye CB",
-      title: "Software Developer",
-      socials: [
-        { twitter: "twitter/meyeneudo", github: "facebook/meyeneudoh" },
-      ],
-    },
-    {
-      id: 6,
-      image:
-        "https://res.cloudinary.com/gfellah45/image/upload/v1601374966/img/pictures_tdnnso.png",
-      name: "Okoye CB",
-      title: "Software Developer",
-      socials: [
-        { twitter: "twitter/meyeneudo", github: "facebook/meyeneudoh" },
-      ],
-    },
-  ]
   return (
-    <div>
+    <div className="talent">
       <DefaultLayout>
         {/* sample view */}
         <Hero title={"Talent"} />
         <div className="container text-center mt-6 mb-6">
-          <h3 className="title pt-4">Our Previous Candidates</h3>
+          <h3 className="talent__title pt-4">Our Previous Candidates</h3>
           <div className="row mt-4">
             {data.map(candidate => (
               <div className="col-12 col-md-6 col-lg-4 " key={candidate.id}>
@@ -90,7 +25,7 @@ function Talent() {
                 />
               </div>
             ))}
-            <div className="button__container ">
+            <div className="talent__button__container ">
               <div className="d-flex justify-content-center">
                 <button className="button--outline m-2">
                   {" "}
